+++ bee-storage/bee-storage-memory/src/access/fetch.rs
@@ -12,7 +12,6 @@ use bee_message::{
     address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
-    payload::indexation::PaddedIndex,
     Message, MessageId,
 };
 use bee_storage::{access::Fetch, backend::StorageBackend, system::System};
@@ -34,7 +33,6 @@ impl_fetch!(u8, System, system);
 impl_fetch!(MessageId, Message, message_id_to_message);
 impl_fetch!(MessageId, MessageMetadata, message_id_to_metadata);
 impl_fetch!(MessageId, Vec<MessageId>, message_id_to_message_id);
-impl_fetch!(PaddedIndex, Vec<MessageId>, index_to_message_id);
 impl_fetch!(OutputId, CreatedOutput, output_id_to_created_output);
 impl_fetch!(OutputId, ConsumedOutput, output_id_to_consumed_output);
 impl_fetch!(Ed25519Address, Vec<OutputId>, ed25519_address_to_output_id);