+++ bee-storage/bee-storage-rocksdb/src/access/exist.rs
@@ -14,7 +14,6 @@ use bee_message::{
     address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
-    payload::indexation::PaddedIndex,
     Message, MessageId,
 };
 use bee_storage::access::Exist;
@@ -50,18 +49,6 @@ impl Exist<(MessageId, MessageId), ()> for Storage {
         Ok(self
             .inner
             .get_cf(self.cf_handle(CF_MESSAGE_ID_TO_MESSAGE_ID)?, key)?
-            .is_some())
-    }
-}
-
-impl Exist<(PaddedIndex, MessageId), ()> for Storage {
-    fn exist(&self, (index, message_id): &(PaddedIndex, MessageId)) -> Result<bool, <Self as StorageBackend>::Error> {
-        let mut key = index.as_ref().to_vec();
-        key.extend_from_slice(message_id.as_ref());
-
-        Ok(self
-            .inner
-            .get_cf(self.cf_handle(CF_INDEX_TO_MESSAGE_ID)?, key)?
             .is_some())
     }
 }