+++ bee-storage/bee-storage-rocksdb/src/access/fetch.rs
@@ -13,7 +13,6 @@ use bee_message::{
     address::Ed25519Address,
     milestone::{Milestone, MilestoneIndex},
     output::OutputId,
-    payload::indexation::PaddedIndex,
     Message, MessageId,
 };
 use bee_storage::{access::Fetch, system::System};
@@ -65,23 +64,6 @@ impl Fetch<MessageId, Vec<MessageId>> for Storage {
                     MessageId::from(child)
                 })
                 .take(self.config.fetch_edge_limit)
-                .collect(),
-        ))
-    }
-}
-
-impl Fetch<PaddedIndex, Vec<MessageId>> for Storage {
-    fn fetch(&self, index: &PaddedIndex) -> Result<Option<Vec<MessageId>>, <Self as StorageBackend>::Error> {
-        Ok(Some(
-            self.inner
-                .prefix_iterator_cf(self.cf_handle(CF_INDEX_TO_MESSAGE_ID)?, index)
-                .map(|(key, _)| {
-                    let (_, message_id) = key.split_at(PaddedIndex::LENGTH);
-                    // Unpacking from storage is fine.
-                    let message_id: [u8; MessageId::LENGTH] = message_id.try_into().unwrap();
-                    MessageId::from(message_id)
-                })
-                .take(self.config.fetch_index_limit)
                 .collect(),
         ))
     }