+++ bee-storage/bee-storage-sled/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 - All `Balance` operations;
+- All `PaddedIndex` operations;
 
 ## 0.5.0 - 2022-XX-XX